requirements.txt
boto3
PyJWT
pytest